+++ s3torchconnectorclient/python/tst/integration/test_logging.py
@@ -2,7 +2,7 @@
 #  // SPDX-License-Identifier: BSD
 import os.path
 import sys
-from datetime import datetime
+import tempfile
 
 import pytest
 
@@ -11,9 +11,10 @@ import logging
 import os
 import sys
 
-os.environ["ENABLE_CRT_LOGS"] = "{0}"
-
-logs_dir_path = "{1}"
+s3_uri = sys.argv[1]
+region = sys.argv[2]
+os.environ["ENABLE_CRT_LOGS"] = sys.argv[3]
+logs_dir_path = sys.argv[4]
 if logs_dir_path != "":
     os.environ["CRT_LOGS_DIR_PATH"] = logs_dir_path
 
@@ -25,10 +26,6 @@ logging.basicConfig(
 )
 logging.getLogger().setLevel(logging.INFO)
 
-if len(sys.argv) != 3:
-    exit("The script needs an S3 uri and a region")
-s3_uri = sys.argv[1]
-region = sys.argv[2]
 map_dataset = S3MapDataset.from_prefix(s3_uri, region=region)
 obj = map_dataset[0]
 assert obj is not None
@@ -68,6 +65,7 @@ import subprocess
             # Python log level is set to INFO in the test script
             ["TRACE s3torchconnector.s3map_dataset"],
         ),
+        ("OFF", ["INFO s3torchconnector.s3map_dataset"], ["awscrt"]),
     ],
 )
 def test_logging_valid(log_level, should_contain, should_not_contain, image_directory):
@@ -78,36 +76,24 @@ def test_logging_valid(log_level, should_contain, should_not_contain, image_dire
 
 
 def test_logging_to_file(image_directory):
-    dt_string = datetime.now().strftime("%d-%m-%Y-%H-%M-%S.%f")
-    log_dir = f"/tmp/test-logs-{dt_string}/"
-    out, err = _start_subprocess("INFO", image_directory, log_dir)
-    # Standard output contains Python output
-    assert err == ""
-    assert "INFO s3torchconnector.s3map_dataset" in out
-    assert "awscrt" not in out
-    # Verifying file logging
-    assert os.path.exists(log_dir)
-    files = os.listdir(log_dir)
-    assert len(files) == 1
-    log_file = f"{log_dir}{files[0]}"
-    assert os.path.isfile(log_file)
-    f = open(log_file, "r")
-    file_content = f.read()
-    assert all(s in file_content for s in ["awscrt", "INFO"])
-    assert all(
-        s not in file_content
-        for s in ["INFO s3torchconnector.s3map_dataset", "DEBUG", "TRACE"]
-    )
-    # Cleanup logs file and directory
-    os.remove(log_file)
-    os.rmdir(log_dir)
-
-
-def test_logging_off(image_directory):
-    out, err = _start_subprocess("off", image_directory)
-    assert err == ""
-    assert "INFO s3torchconnector.s3map_dataset" in out
-    assert "awscrt" not in out
+    with tempfile.TemporaryDirectory() as log_dir:
+        print("Created temporary directory", log_dir)
+        out, err = _start_subprocess("INFO", image_directory, log_dir)
+        # Standard output contains Python output
+        assert err == ""
+        assert "INFO s3torchconnector.s3map_dataset" in out
+        assert "awscrt" not in out
+        files = os.listdir(log_dir)
+        assert len(files) == 1
+        log_file = os.path.join(log_dir, files[0])
+        assert os.path.isfile(log_file)
+        with open(log_file) as f:
+            file_content = f.read()
+            assert all(s in file_content for s in ["awscrt", "INFO"])
+            assert all(
+                s not in file_content
+                for s in ["INFO s3torchconnector.s3map_dataset", "DEBUG", "TRACE"]
+            )
 
 
 def _start_subprocess(log_level, image_directory, logs_directory=""):
@@ -115,9 +101,11 @@ def _start_subprocess(log_level, image_directory, logs_directory=""):
         [
             sys.executable,
             "-c",
-            PYTHON_TEST_CODE.format(log_level, logs_directory),
+            PYTHON_TEST_CODE,
             image_directory.s3_uri,
             image_directory.region,
+            log_level,
+            logs_directory,
         ],
         stdout=subprocess.PIPE,
         stderr=subprocess.PIPE,