+++ s3torchconnector/src/s3torchconnector/_s3client/_mock_s3client.py
@@ -7,6 +7,7 @@ from s3torchconnectorclient._mountpoint_s3_client import (
 )
 
 from . import S3Client
+from .._user_agent import UserAgent
 
 """
 _mock_s3client.py
@@ -15,9 +16,20 @@ _mock_s3client.py
 
 
 class MockS3Client(S3Client):
-    def __init__(self, region: str, bucket: str, part_size: int = 8 * 1024 * 1024):
-        super().__init__(region)
-        self._mock_client = MockMountpointS3Client(region, bucket, part_size=part_size)
+    def __init__(
+        self,
+        region: str,
+        bucket: str,
+        part_size: int = 8 * 1024 * 1024,
+        user_agent: UserAgent = None,
+    ):
+        super().__init__(region, user_agent=user_agent)
+        self._mock_client = MockMountpointS3Client(
+            region,
+            bucket,
+            part_size=part_size,
+            user_agent_prefix=self.user_agent_prefix,
+        )
 
     def add_object(self, key: str, data: bytes) -> None:
         self._mock_client.add_object(key, data)